+++ client/src/components/Cards/CardPlan/index.js
@@ -1,0 +1,1 @@
+export { default as CardPlan } from './CardPlan';
