+++ client/src/components/Cards/index.js
@@ -1,3 +1,4 @@
 export * from './CardsFeedback';
 export * from './CardAlert';
 export * from './CardCookie';
+export * from './CardPlan';
