--- client/src/components/Playlist/Playlist.jsx
@@ -1,9 +1,0 @@
-const Playlist = () => {
-  return (
-    <div>
-      <h1>Playlist</h1>
-    </div>
-  );
-};
-
-export default Playlist;
